lock Poetry version in build
I'm explicitly calling out the version of Poetry to be used during
builds to keep that process as stable and consistent as possible.
Though, as far as I know, Poetry doesn't perform any sort of
enforcement or validation around what's in `build-system.requires`
in `pyproject.toml` but it may one day, and as a worst-case, this
acts as documentation.